% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_code.R
\name{Day_of_Year}
\alias{Day_of_Year}
\title{Day of Year}
\usage{
Day_of_Year(day, month, year)
}
\arguments{
\item{day}{Day of the respective date.}

\item{month}{Numeric value from 1 to 12 representing the month of the respective date.}

\item{year}{Year of the respective date.}
}
\value{
The number of the day in the given year.
}
\description{
By specifying a date with day, month and year it returns the number of the day in that respective year.
}
\examples{
Day_of_Year(7,9,1822)

}
