% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_code.R
\name{Daylight_Hours}
\alias{Daylight_Hours}
\title{Daylight Hours}
\usage{
Daylight_Hours(sunset_hour_angle)
}
\arguments{
\item{sunset_hour_angle}{Hour angle at sunset in radians.}
}
\value{
Maximum sunshine duration.
}
\description{
Calculate maximum sunshine duration from sunset hour angle.
}
\examples{
Daylight_Hours(sunset_hour_angle(latitude_radians(22,45,north=TRUE),solar_decimation(07,09,1882)))
Daylight_Hours(1.60969)
}
