% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_code.R
\name{ea_dewpoint}
\alias{ea_dewpoint}
\title{Actual Vapor Pressure (dew point)}
\usage{
ea_dewpoint(Tdew)
}
\arguments{
\item{Tdew}{Dew point temperature in Celsius.}
}
\value{
Actual vapor pressure in kPa.
}
\description{
Returns the actual vapor pressure in kPa considering the dew point.
}
\examples{
ea_dewpoint(20)

}
