% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_code.R
\name{ea_RH}
\alias{ea_RH}
\title{Actual Vapor Pressure (relative humidity)}
\usage{
ea_RH(Tmax, Tmin, RHmax, RHmin)
}
\arguments{
\item{Tmax}{Maximum temperature in Celcius.}

\item{Tmin}{Minimum temperature in Celcius.}

\item{RHmax}{Maximum relative humidity.}

\item{RHmin}{Minimim relative humidity.}
}
\value{
Actual vapor pressure in kPa.
}
\description{
Estimate the actual vapor pressure using relative humidity.
}
\examples{
ea_RH(25,32,100,80)

}
