% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_code.R
\name{ea_RHmax}
\alias{ea_RHmax}
\title{Actual Vapor Pressure (maximum relative humidity)}
\usage{
ea_RHmax(Tmin, RHmax)
}
\arguments{
\item{Tmin}{Minimum temperature in Celcius.}

\item{RHmax}{Maximum relative humidity.}
}
\value{
Actual vapor pressure in kPa.
}
\description{
Estimate the actual vapor pressure with only maximum relative humidity.
}
\examples{
ea_RHmax(25,100)

}
