% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_code.R
\name{ea_RHmean}
\alias{ea_RHmean}
\title{Actual Vapor Pressure (mean relative humidity)}
\usage{
ea_RHmean(Temperatures, RHmean)
}
\arguments{
\item{Temperatures}{Single temperature or a vector of temperature in Celcius.}

\item{RHmean}{Mean relative humidity.}
}
\value{
Actual vapor pressure in kPa.
}
\description{
Estimate the actual vapor pressure with only mean relative humidity.
}
\note{
The FAO56 recommendation is to use the mean vapor saturation pressure of the values estimated at the maximum and minimum temperature in a day instead of measuring only at mean temperature. As the following actual vapor pressure estimation requires the estimation of vapor saturation pressure, to satisfy both uses (with mean daily temperature and with maximum and minimum temperatures), which are common, and cases that evapotranspiration is estimated for longer periods, this function was defined on such way.
}
\examples{
ea_RHmean(c(25,32),90)

}
