% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_code.R
\name{es}
\alias{es}
\title{Mean Vapor Saturation Pressure}
\usage{
es(temperatures)
}
\arguments{
\item{temperatures}{A single temperature or a vector of temperatures in Celsius.}
}
\value{
Mean vapor saturation pressure in kPa across the given temperatures.
}
\description{
Returns the mean vapor saturation pressure in kPa across a set of temperatures in Celsius.
}
\note{
The FAO56 recommendation is to use the mean vapor saturation pressure of the values estimated at the maximum and minimum temperature in a day instead of measuring only at mean temperature. To satisfy both uses (with mean daily temperature and with maximum and minimum temperatures), which are common, and cases that evapotranspiration is estimated for longer periods, the following function was created.
}
\examples{
es(c(20,25))

}
