% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_code.R
\name{esT}
\alias{esT}
\title{Vapor Saturation Pressure}
\usage{
esT(Temperature)
}
\arguments{
\item{Temperature}{Temperature in Celsius.}
}
\value{
vapor saturation pressure in kPa.
}
\description{
Returns the vapor saturation pressure in kPa for a given temperature in Celsius.
}
\examples{
esT(20)

}
