% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_code.R
\name{inverse_distance_earth_sun}
\alias{inverse_distance_earth_sun}
\title{Inverse Relative Distance from Earth to Sun}
\usage{
inverse_distance_earth_sun(day, month, year, day_of_year)
}
\arguments{
\item{day}{Day of the respective date.}

\item{month}{Numeric value from 1 to 12 representing the month of the respective date.}

\item{year}{Year of the respective date.}

\item{day_of_year}{The number of the day in a given year.}
}
\value{
The inverse relative distance from earth to sun in a given day of the year.
}
\description{
By specifying the date with day, month and year or the day of the year it returns the inverse relative distance from earth to sun in the respective day of the year.
}
\examples{
inverse_distance_earth_sun(7,9,1822)
inverse_distance_earth_sun(day_of_year=250)
}
