% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_code.R
\name{latitude_radians}
\alias{latitude_radians}
\title{Latitude in Radians}
\usage{
latitude_radians(degrees, minutes, north = TRUE)
}
\arguments{
\item{degrees}{Degrees of the latitude.}

\item{minutes}{Minutes of the latitude.}

\item{north}{Boolean indicating if the refeered location is in the north hemisphere. Default is north=TRUE.}
}
\value{
Latitude in radians.
}
\description{
By specifying the degrees and minutes of a latitude, and specifying if the refeered location is in or not in the north hemisphere it converts the latitude from degrees and minutes to radians.
}
\examples{
latitude_radians(22,45)
latitude_radians(22,45,FALSE)
latitude_radians(22,45,TRUE)
}
