% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_code.R
\name{Patmz}
\alias{Patmz}
\title{Atmospheric Pressure}
\usage{
Patmz(z)
}
\arguments{
\item{z}{Elevation of the location}
}
\value{
The estimated atmospheric pressure in kPa on the location at elevation z.
}
\description{
Estimate atmospheric pressure based on the elevation z of a location.
}
\examples{
Patmz(100)

}
