% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_code.R
\name{psychrometric}
\alias{psychrometric}
\title{Psychrometric Coefficient}
\usage{
psychrometric(Patm)
}
\arguments{
\item{Patm}{Atmospheric pressure in kPa.}
}
\value{
Psychrometric coefficient at the location.
}
\description{
Returns the psychrometric coefficient based on the atmospheric pressure in a location.
}
\examples{
psychrometric(100)

}
