% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_code.R
\name{Ra}
\alias{Ra}
\title{Extraterresterial Radiation}
\usage{
Ra(
  day,
  month,
  year,
  lat_degrees,
  lat_minutes,
  north = TRUE,
  day_of_year,
  lat_radians
)
}
\arguments{
\item{day}{Day of the respective date.}

\item{month}{Numeric value from 1 to 12 representing the month of the respective date.}

\item{year}{Year of the respective date.}

\item{lat_degrees}{Degrees of the latitude.}

\item{lat_minutes}{Minutes of the latitude.}

\item{north}{Boolean indicating if the refeered location is in north hemisphere. Default is north=TRUE.}

\item{day_of_year}{The number of the day in a given year.}

\item{lat_radians}{Latitude in radians.}
}
\value{
Extraterresterial radiation.
}
\description{
Calculate extraterresterial radiation in a specific day of year and latitude.
}
\note{
It is only necessary to specify the day, month and year of the day or the respective day_of_year, if both are specified day_of_year will be used. Also, it is only necessary to specify lat_degrees and lat_minutes or lat_radians, if both are specified lat_radians will be used.
}
\examples{
Ra(7,9,1822,22,45,TRUE)
Ra(day_of_year=250,lat_radians=0.3970624)
}
