% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_code.R
\name{Rn}
\alias{Rn}
\title{Net Radiation}
\usage{
Rn(Rns, Rnl)
}
\arguments{
\item{Rns}{Net shortwave radiation.}

\item{Rnl}{Net longwave radiation.}
}
\value{
Net radiation.
}
\description{
Calculate the difference among incoming shortwave radiation and outgoing longwave radiation.
}
\examples{
Rn(26,9e-7)
}
