% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_code.R
\name{Rnl}
\alias{Rnl}
\title{Net Longwave Radiation (Rnl)}
\usage{
Rnl(Tmax, Tmin, ea, Rs, Rso)
}
\arguments{
\item{Tmax}{Maximum temperature in Celsius.}

\item{Tmin}{Minimum temperature in Celsius.}

\item{ea}{Actual vapor pressure.}

\item{Rs}{Incoming solar radiation.}

\item{Rso}{Incoming solar radiation at clean sky.}
}
\value{
Net solar longwave radiation.
}
\description{
Calculate net outcoming longwave radiation from maximum and minimum temperatures, actual vapor pressure, incoming solar radiation and incoming solar radiation at clean sky.
}
\examples{
Rnl(30,22,2.410449,27,30)
}
