% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_code.R
\name{Rns}
\alias{Rns}
\title{Net Solar or Net Shortwave Radiation}
\usage{
Rns(Rs, alpha = 0.23)
}
\arguments{
\item{Rs}{Incoming solar radiation.}

\item{alpha}{Albedo. Default value is 0.23, representing the reference grass crop.}
}
\value{
Net solar (shortwave) radiation resulting from the balance among incoming and reflected solar radiation.
}
\description{
Calculate net solar (shortwave) radiation resulting from the balance among incoming and reflected radiation using albedo and the estimated incoming solar radiation.
}
\examples{
Rns(Rs_Angstrom(12,12.79,Ra(7,9,1882,22,45,TRUE)))
Rns(32)
Rns(Rs_Angstrom(12,12.79,Ra(7,9,1882,22,45,TRUE)),alpha=0.17)
Rns(32,alpha=0.17)
}
