% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_code.R
\name{Rs_Angstrom}
\alias{Rs_Angstrom}
\title{Angstrom Formula}
\usage{
Rs_Angstrom(Sunshine_Duration, Daylight_Hours, Ra, as = 0.25, bs = 0.5)
}
\arguments{
\item{Sunshine_Duration}{Actual sunshine duration in hours.}

\item{Daylight_Hours}{Maximum sunshine duration in hours.}

\item{Ra}{Extraterresterial radiation.}

\item{as}{Fraction of extraterresterial radiation reaching the earth on overcast days.}

\item{bs}{as + bs is the fraction of extraterresterial radiation reaching earth on clear days.}
}
\value{
Solar radiation calculated from Angstrom formula.
}
\description{
Calculate solar radiation using Angstrom formula using the fraction of radiation reaching the earth in overcast and clear days, and the actual and potential sunshine hours.
}
\examples{
Rs_Angstrom(12,12.29713,36.16717)
}
