% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_code.R
\name{Rso}
\alias{Rso}
\title{Solar Radiation on Clean Sky}
\usage{
Rso(Ra, z, as, bs)
}
\arguments{
\item{Ra}{Extraterresterial radiation.}

\item{z}{Elevation above sea level of weather station.}

\item{as}{Fraction of extraterrestrial radiation reaching the earth on overcast days.}

\item{bs}{as + bs is the fraction of extraterrestrial radiation reaching earth on clear days.}
}
\value{
Solar radiation calculated from Angstrom formula.
}
\description{
Calculate solar radiation on clean sky using extraterresterial radiation, elevation above sea level when as and bs are not provided or with as and bs values.
}
\examples{
Rso(Ra(7,9,1822,22,45,TRUE),0)
Rso(Ra(7,9,1822,22,45,TRUE),100)
Rso(36.16717,as=0.3,bs=0.55)
}
