% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_code.R
\name{set_gamma_psy}
\alias{set_gamma_psy}
\title{Psychrometer Psychrometric Constant}
\usage{
set_gamma_psy(a_psy = "ventilated", Patm)
}
\arguments{
\item{a_psy}{Must be one of the following charaters: "ventilated", indicating that the psychrometer is ventilated (Asmann type), with an air movement of some 5 m/s (constant value 0.000662); "natural" for natural ventilated psychrometers (about 1 m/s, constant value 0.000800); "non-ventilated" for non-ventilated psychrometers installed indoors (constant of 0.001200); or a numeric indicating the constant of the psychrometer of the weather station.}

\item{Patm}{Atmospheric pressure in kPa.}
}
\value{
Psychrometric constant of the psychrometer at a given atmospheric pressure.
}
\description{
Set the psychrometric constant of the psychrometer based upon its ventilation method (ventilated, natural, non-ventilated) or according a value of the instrument constant set by the user and the local atmospheric pressure.
}
\examples{
set_gamma_psy(Patm=100)
set_gamma_psy("ventilated",100)
set_gamma_psy("natural",100)
set_gamma_psy("non-ventilated",100)
set_gamma_psy(0.000700,100)
}
