% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_code.R
\name{slope}
\alias{slope}
\title{Vapor Saturation Curve Slope}
\usage{
slope(Temperature)
}
\arguments{
\item{Temperature}{Temperature in Celsius.}
}
\value{
Slope of vapor saturation pressure at a given temperature.
}
\description{
Returns the slope of the vapor saturation curve at a given temperature in Celsius.
}
\examples{
slope(20)

}
