% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_code.R
\name{solar_decimation}
\alias{solar_decimation}
\title{Solar Decimation}
\usage{
solar_decimation(day, month, year, day_of_year)
}
\arguments{
\item{day}{Day of the respective date.}

\item{month}{Numeric value from 1 to 12 representing the month of the respective date.}

\item{year}{Year of the respective date.}

\item{day_of_year}{The number of the day in a given year.}
}
\value{
The solar decimation in radians in a given day of the year.
}
\description{
By specifying the date with day, month and year or the day of the year it returns the solar decimation in the given day of the year.
}
\examples{
solar_decimation(7,9,1822)
solar_decimation(day_of_year=250)
}
