% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_code.R
\name{sunset_hour_angle}
\alias{sunset_hour_angle}
\title{Sunset Hour Angle}
\usage{
sunset_hour_angle(lat_radians, solar_decimation)
}
\arguments{
\item{lat_radians}{Latitude in radians.}

\item{solar_decimation}{Solar decimation in radians.}
}
\value{
Hour angle at sunset.
}
\description{
Returns sunset hour angle by solar decimation and latitude in radians.
}
\examples{
sunset_hour_angle(0.3970624,0.09246253)
sunset_hour_angle(latitude_radians(22,45,TRUE),solar_decimation(7,9,1822))
}
