% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_code.R
\name{u2}
\alias{u2}
\title{Estimated Wind Velocity at 2 m Height}
\usage{
u2(z, uz)
}
\arguments{
\item{z}{Height in meters where wind velocity is measured.}

\item{uz}{Wind velocity at height z.}
}
\value{
Estimated wind velocity at the height of 2 m
}
\description{
Estimate the wind velocity at 2 m when it is measured at a different height from soil.
}
\examples{
u2(10,5)
}
